+++ spec/cli/pack_spec.rb
@@ -1,7 +1,7 @@
 require "spec_helper"
 require "libgems/format"
 
-describe "bpm build when logged in" do
+describe "bpm pack when logged in" do
   let(:email) { "who@example.com" }
 
   before do
@@ -25,12 +25,12 @@ describe "bpm build when logged in" do
   end
 end
 
-describe "bpm build without logging in" do
+describe "bpm pack without logging in" do
   before do
     goto_home
   end
 
-  it "builds a bpm from a given package.json" do
+  it "pack a bpm from a given package.json" do
     FileUtils.cp_r package_fixture("core-test"), "."
     cd "core-test"
     bpm "pack", "-e", "joe@example.com"
@@ -69,7 +69,7 @@ describe "bpm build without logging in" do
   end
 end
 
-describe "bpm build with an invalid package.json" do
+describe "bpm pack with an invalid package.json" do
   before do
     goto_home
     write_api_key("deadbeef")
@@ -84,3 +84,35 @@ describe "bpm build with an invalid package.json" do
     output.should include("There was a problem parsing package.json")
   end
 end
+
+describe "bpm pack npm-compatible package" do
+  before do
+    FileUtils.cp_r package_fixture("backbone"), "."
+    cd "backbone"
+    bpm "pack", :track_stderr => true and wait
+  end
+
+  it "successfully packs" do
+    exit_status.should be_success
+  end
+
+  it "implies the summary field" do
+    package = LibGems::Format.from_file_by_path("backbone-0.5.1.bpkg")
+    package.spec.summary.should == package.spec.description
+  end
+  
+  it "gets name and version" do
+    package = LibGems::Format.from_file_by_path("backbone-0.5.1.bpkg")
+    package.spec.name.should == "backbone"
+    package.spec.version.should == LibGems::Version.new("0.5.1")
+  end
+
+  it "gets the homepage" do
+    package = LibGems::Format.from_file_by_path("backbone-0.5.1.bpkg")
+    package.spec.homepage.should == 'http://documentcloud.github.com/backbone/'
+  end
+  
+    
+end
+
+